next_dim/requirements.txt
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
SpeechRecognition
PyAudio
gTTS
playsound==1.2.2